+++ yaml-config-updater/src/main/java/ru/vyarus/yaml/updater/parse/comments/CommentsReader.java
@@ -231,7 +231,7 @@ public final class CommentsReader {
 
         @Override
         public String toString() {
-            return key + ": " + value;
+            return key + ":" + value;
         }
     }